refactor: all ads are ts files now
