+++ src/types/AdsConsent.interface.ts
@@ -321,19 +321,6 @@ export interface AdsConsentFormOptions {
 }
 
 /**
- * A `AdShowOptions` interface used when showing an ad.
- */
-export interface AdShowOptions {
-  /**
-   * - On Android, enables [immersive mode](https://developer.android.com/training/system-ui/immersive).
-   * - On iOS, this has no effect on how the ad is shown.
-   *
-   * @android
-   */
-  immersiveModeEnabled?: boolean;
-}
-
-/**
  * The result of a Google-rendered consent form.
  */
 export interface AdsConsentFormResult {
