+++ src/types/RequestOptions.ts
@@ -13,21 +13,6 @@ export interface ServerSideVerificationOptions {
 }
 
 /**
- * A `RewardedAdReward` returned from rewarded ads.
- */
-export interface RewardedAdReward {
-  /**
-   * The reward name, e.g. 'coins', 'diamonds'.
-   */
-  type: string;
-
-  /**
-   * The number value of the reward, e.g. 10
-   */
-  amount: number;
-}
-
-/**
  * The `RequestOptions` interface. Used when passing additional request options before an advert is loaded.
  */
 export interface RequestOptions {
